+++ src/p_setup.cpp
@@ -1318,7 +1318,7 @@ void P_LoadSectors (MapData * map)
 		ss->ceilingplane.ic = -FRACUNIT;
 		SetTexture(ss, i, sector_t::floor, ms->floorpic);
 		SetTexture(ss, i, sector_t::ceiling, ms->ceilingpic);
-		ss->lightlevel = clamp (LittleShort(ms->lightlevel), (short)0, (short)255);
+		ss->lightlevel = (BYTE)clamp (LittleShort(ms->lightlevel), (short)0, (short)255);
 		if (map->HasBehavior)
 			ss->special = LittleShort(ms->special);
 		else	// [RH] Translate to new sector special
@@ -1743,7 +1743,7 @@ void P_FinishLoadingLineDef(line_t *ld, int alpha)
 	ld->backsector  = ld->sidenum[1]!=NO_SIDE ? sides[ld->sidenum[1]].sector : 0;
 	float dx = FIXED2FLOAT(ld->v2->x - ld->v1->x);
 	float dy = FIXED2FLOAT(ld->v2->y - ld->v1->y);
-	int linenum = ld-lines;
+	int linenum = int(ld-lines);
 
 	if (ld->frontsector == NULL)
 	{
@@ -2058,7 +2058,7 @@ static void P_LoopSidedefs ()
 		// as their left edge.
 		line_t *line = &lines[sides[i].linenum];
 		int lineside = (line->sidenum[0] != (DWORD)i);
-		int vert = (lineside ? line->v2 : line->v1) - vertexes;
+		int vert = int((lineside ? line->v2 : line->v1) - vertexes);
 		
 		sidetemp[i].b.lineside = lineside;
 		sidetemp[i].b.next = sidetemp[vert].b.first;
@@ -2088,18 +2088,18 @@ static void P_LoopSidedefs ()
 		{
 			if (sidetemp[i].b.lineside)
 			{
-				right = line->v1 - vertexes;
+				right = int(line->v1 - vertexes);
 			}
 			else
 			{
-				right = line->v2 - vertexes;
+				right = int(line->v2 - vertexes);
 			}
 
 			right = sidetemp[right].b.first;
 
 			if (right == NO_SIDE)
 			{ // There is no right side!
-				Printf ("Line %d's right edge is unconnected\n", linemap[line-lines]);
+				Printf ("Line %d's right edge is unconnected\n", linemap[unsigned(line-lines)]);
 				continue;
 			}
